+++ src/pages/Books/Books.jsx
@@ -3,7 +3,7 @@ import { useLocation } from "react-router-dom";
 import "./Books.css";
 import Breadcrumb from "../../components/Breadcrumb/Breadcrumb";
 import PublicationList from "../../components/PublicationList/PublicationList";
-import { getBooks } from "../../data/BooksInfo";
+import { getBooks } from "../../data/Books";
 
 function Books() {
   const location = useLocation();
