+++ src/pages/Zines/Zines.jsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 import "./Zines.css";
 import Breadcrumb from "../../components/Breadcrumb/Breadcrumb";
-import { getZines } from "../../data/ZinesInfo"
+import { getZines } from "../../data/Zines"
 import PublicationList from "../../components/PublicationList/PublicationList";
 
 function Zines() {
